+++ docs/CHANGELOG.md
@@ -6,3 +6,4 @@ Changelog
 - Init: Initialize Module "Text editor"
 - Fix #5: Position of "Edit" Handler
 - Enh #2: Create new text files
+- Enh #1: Store file version on update content